added separate files for potentials and utility functions
