+++ potentials.h
@@ -1,0 +1,2 @@
+double getLennardJonesEnergy(double distance);
+double getLennardJonesForce(double distance);
