+++ utility.h
@@ -1,0 +1,5 @@
+#include <string>
+
+// checks if char array represents a number
+bool isNumber(char number[]);
+bool fileExists(const std::string &name);
